requirements.txt
streamlit==1.29.0
chromadb==0.4.22
sentence-transformers==2.2.2
anthropic==0.18.0
PyMuPDF==1.23.0
camelot-py[cv]==0.11.0
pandas==2.1.0
rank-bm25==0.2.2
tiktoken==0.5.1
numpy==1.24.3
python-dotenv==1.0.0

